Added persistence for GameAssets

+++ src/main/java/de/hsrm/mi/swt02/backend/api/map/dto/AddGameAssetRequestDTO.java
@@ -2,8 +2,8 @@ package de.hsrm.mi.swt02.backend.api.map.dto;
 
 public record AddGameAssetRequestDTO(
         int objectTypeId,
-        int x,
-        int y,
+        double x,
+        double y,
         int rotation,
         String texture
 ) {}
